+++ model_compression_toolkit/core/common/pruning/graph_utils/pruning_section_mask.py
@@ -1,0 +1,16 @@
+import numpy as np
+class PruningSectionMask:
+    """
+    Contains the masks for the convolution nodes in a pruning section, which
+    indicate which output channels remain and which are pruned.
+    """
+
+    def __init__(self,
+                 first_node_input_mask: np.ndarray,
+                 first_node_output_mask: np.ndarray,
+                 second_node_output_mask: np.ndarray
+                 ):
+        self.first_node_input_mask = first_node_input_mask
+        self.first_node_output_mask = first_node_output_mask
+        self.second_node_output_mask = second_node_output_mask
+
